move scores tests to own file
